BUG: fixed the bug with python 3 ( dict.iteritem() not working) API moved the circular_average function to core.py and the test to test_core.py
