+++ skxray/tests/test_core.py
@@ -46,6 +46,7 @@ import sys
 from nose.tools import assert_equal, assert_true, raises
 
 import skxray.core as core
+import skxray.roi as roi
 
 from skxray.testing.decorators import known_fail_if
 import numpy.testing as npt
@@ -563,6 +564,22 @@ def test_radial_grid():
     assert_equal(a[3, 3], 0)
     assert_equal(a[3, 4], 1)
 
+def test_circular_average():
+    image = np.zeros((12,12))
+    calib_center = (5, 5)
+    inner_radius = 1
+
+    edges = roi.ring_edges(inner_radius, width=1, spacing=1, num_rings=2)
+    labels = roi.rings(edges, calib_center, image.shape)
+    image[labels==1] = 10
+    image[labels==2] = 10
+    bin_cen, ring_avg = core.circular_average(image, calib_center, nx=6)
+
+    assert_array_almost_equal(bin_cen, [0.70710678, 2.12132034,
+                                        3.53553391,  4.94974747,  6.36396103,
+                                        7.77817459], decimal=6)
+    assert_array_almost_equal(ring_avg, [8., 2.5, 5.55555556, 0.,
+                                         0., 0.], decimal=6)
 
 if __name__ == '__main__':
     import nose
